requirements.txt
python-telegram-bot>=20.0
python-dotenv
pyyaml
requests
httpx
pillow
textual>=0.41.0
rich>=13.0.0
solana>=0.30.0
solders>=0.18.0
